requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
flask-cors==4.0.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
Jinja2==3.1.2
Werkzeug==2.3.7
